+++ CHANGELOG.md
@@ -33,6 +33,7 @@ Test/interim releases are not documented here.
 ### Add
 - Add convenience helpers for `SmTransformer`.
 - Add diagram based toml config for `RenderConfig` and *some* `SmRunner.Settings`
+  - Useful for plantuml as well which previously never had diagram based render config support.
   - https://github.com/StateSmith/StateSmith/issues/335
 
 ---